Add the sessionId parameter to the endpoint function
Add the sessionId optional parameter to the setLastDispatchedRequestId and the getter

Add the sessionId parameter so middleware update of dispatched requests

Remove the requestsLog section of beccaccino redux state.

Add request log getter for SessionManager

Refactored selectors in order to handle the new session management.

wip

Fix tests for requestsLog in redux state

Fix a test for beccaccinoSelector

  requestsMetadata section of the state for the selector input
  was outside the beccaccino section of the state